bugfix to update_repo

+++ scripts/update_repo.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 cd ~/raspberry_pi
-git stash 
-git reset --hard HEAD
+git status | grep modified >/dev/null
+LC=$?
+if [ $LC -eq 0 ]; then
+    git stash 
+    git reset --hard HEAD
+fi
 git pull
-git stash pop
+if [ $LC -eq 0 ]; then
+    git stash pop
+fi
 
 ~/raspberry_pi/scripts/post_update_setup.sh
